Keep runtime types independent of server implementation
